# schemas/v1/openapi.json
{
  "components": {
    "schemas": {
...
      "PredictionRequest": {
        "additionalProperties": false,
        "description": "Closed request shared by synchronous prediction and reference solve.",
        "properties": {
          "reynolds": {
            "maximum": 300.0,
            "minimum": 40.0,
            "title": "Reynolds",
            "type": "number"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "shape": {
            "$ref": "#/components/schemas/ShapeRequest"
          }
        },
        "required": [
          "shape",
          "reynolds"
        ],
        "title": "PredictionRequest",
        "type": "object"
      },
      "PredictionResponse": {
        "additionalProperties": false,
        "description": "Identity-rich, bounded synchronous prediction response.",
        "properties": {
          "case_id": {
            "pattern": "^[0-9a-f]{20}$",
            "title": "Case Id",
            "type": "string"
          },
          "cd_field": {
            "title": "Cd Field",
            "type": "number"
          },
          "cd_head": {
            "title": "Cd Head",
            "type": "number"
          },
          "consistency": {
            "$ref": "#/components/schemas/ConsistencyFlags"
          },
          "correlation_id": {
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
            "title": "Correlation Id",
            "type": "string"
          },
          "dataset_id": {
            "pattern": "^[0-9a-f]{20}$",
            "title": "Dataset Id",
            "type": "string"
          },
          "fields_npz": {
            "$ref": "#/components/schemas/EncodedArtifact"
          },
          "fields_png": {
            "$ref": "#/components/schemas/EncodedArtifact"
          },
          "inference_ms": {
            "minimum": 0.0,
            "title": "Inference Ms",
            "type": "number"
          },
          "model_id": {
            "pattern": "^[0-9a-f]{20}$",
            "title": "Model Id",
            "type": "string"
          },
          "report_id": {
            "pattern": "^[0-9a-f]{20}$",
            "title": "Report Id",
            "type": "string"
          },
          "request_ms": {
            "minimum": 0.0,
            "title": "Request Ms",
            "type": "number"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "validation_status": {
            "enum": [
              "green",
              "red"
            ],
            "title": "Validation Status",
            "type": "string"
          }
        },
        "required": [
          "correlation_id",
          "case_id",
          "fields_png",
          "fields_npz",
          "cd_head",
          "cd_field",
          "consistency",
          "validation_status",
          "model_id",
          "dataset_id",
          "report_id",
          "inference_ms",
          "request_ms"
        ],
        "title": "PredictionResponse",
        "type": "object"
      },
...
      "PublicErrorResponse": {
        "additionalProperties": false,
        "description": "Envelope shared by every non-success HTTP response.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/PublicError"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "error"
        ],
        "title": "PublicErrorResponse",
        "type": "object"
      },
...
      "SolveAccepted": {
        "additionalProperties": false,
        "description": "Accepted reference-solve job and its two retrieval paths.",
        "properties": {
          "case_id": {
            "pattern": "^[0-9a-f]{20}$",
            "title": "Case Id",
            "type": "string"
          },
          "events_url": {
            "maxLength": 50,
            "minLength": 50,
            "title": "Events Url",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "job_id": {
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
            "title": "Job Id",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "state": {
            "const": "queued",
            "title": "State",
            "type": "string"
          },
          "status_url": {
            "maxLength": 43,
            "minLength": 43,
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "case_id",
          "state",
          "status_url",
          "events_url",
          "expires_at"
        ],
        "title": "SolveAccepted",
        "type": "object"
      },
...
      "SolveEvent": {
        "additionalProperties": false,
        "description": "Monotonic replayable solve state event.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SolveEventData"
          },
          "event": {
            "enum": [
              "queued",
              "running",
              "progress",
              "completed",
              "failed"
            ],
            "title": "Event",
            "type": "string"
          },
          "job_id": {
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
            "title": "Job Id",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "sequence": {
            "minimum": 1.0,
            "title": "Sequence",
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "sequence",
          "job_id",
          "timestamp",
          "event",
          "data"
        ],
        "title": "SolveEvent",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Version-one educational wind-tunnel API. A red validation report remains visible and serviceable; integrity or identity failures make prediction unavailable.",
    "summary": "Strict prediction and asynchronous reference-solve contracts.",
    "title": "Soufflerie HTTP API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/predict": {
      "post": {
        "operationId": "predict_v1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request body exceeds 16 KiB."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Content type is not JSON."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request violates the closed schema."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Required capacity is unavailable."
          }
        },
        "summary": "Predict fields and consistency for one supported design"
      }
    },
    "/solve": {
      "post": {
        "operationId": "solve_submit_v1",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolveAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Idempotency binding or event replay cursor conflicts with retained state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request body exceeds 16 KiB."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Content type is not JSON."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request violates the closed schema."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Required capacity is unavailable."
          }
        },
        "summary": "Admit one bounded asynchronous reference solve"
      }
    },
    "/solve/{job_id}": {
      "get": {
        "operationId": "solve_status_v1",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 36,
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolveStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Solve job is unknown or expired."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request violates the closed schema."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Required capacity is unavailable."
          }
        },
        "summary": "Poll one retained reference-solve state"
      }
    },
    "/solve/{job_id}/events": {
      "get": {
        "operationId": "solve_events_v1",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 36,
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "x-event-data-schema": {
                    "$ref": "#/components/schemas/SolveEvent"
                  }
                }
              }
            },
            "description": "Replayable SSE frames whose data is one SolveEvent JSON object."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Solve job is unknown or expired."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Idempotency binding or event replay cursor conflicts with retained state."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request violates the closed schema."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Required capacity is unavailable."
          }
        },
        "summary": "Open the replayable server-sent solve event stream"
      }
    }
  }
}